Update README.md and fix bugs introduced to TD3 and DDPG after buffer dtypes update in 5413b5a6e97347535c5067062254cc5d58405077

+++ xagents/ddpg/agent.py
@@ -60,7 +60,7 @@ class DDPG(OffPolicy):
         Returns:
             actions
         """
-        actions = self.actor(tf.numpy_function(self.get_states, [], tf.float32))
+        actions = self.actor(tf.numpy_function(self.get_states, [], tf.float64))
         actions += tf.random.normal(
             shape=(self.n_envs, self.n_actions), stddev=self.step_noise_coef
         )
